+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.10] - 2022-09-12
+### Fixed
+- Fixed a bug in `bordafuse.py`: `get_candidates` raised error if no run had retrieved docs for a given query.
+- Fixed a bug in `borda_norm.py`: `get_candidates` raised error if no run had retrieved docs for a given query.
+- Fixed a bug in `condorcet.py`: `get_candidates` raised error if no run had retrieved docs for a given query.
+
 ## [0.2.9] - 2022-08-29
 ### Fixed
 - Fixed a bug in `report.py:Report`: some metric labels were missing.